requirements.txt
boto
certifi
chardet
cycler
dataclasses
future
idna
kiwisolver
matplotlib>=3.3.1
numpy>=1.19.1
opencv-python
pandas>=1.1.1
Pillow>=7.2.0
pyparsing
pysparkling
python-dateutil
python-json-logger
pytz==2021.1
requests
scipy
six
torch==1.10.2
torchvision==0.11.3
urllib3==1.26.5
openpifpaf==0.12.14
psutil
imutils
numpy
datetime
pykafka==2.0.0